feat: rename lib to src + test from dist
