+++ src/providerRegistry.js
